+++ arduino-ide-extension/src/browser/dialogs/settings/settings-component.tsx
@@ -1,336 +1,22 @@
 import * as React from 'react';
-import { injectable, inject, postConstruct } from 'inversify';
-import { Widget } from '@phosphor/widgets';
-import { Message } from '@phosphor/messaging';
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
 import 'react-tabs/style/react-tabs.css';
 import { Disable } from 'react-disable';
-import URI from '@theia/core/lib/common/uri';
-import { Emitter } from '@theia/core/lib/common/event';
-import { Deferred } from '@theia/core/lib/common/promise-util';
 import { deepClone } from '@theia/core/lib/common/objects';
 import { FileService } from '@theia/filesystem/lib/browser/file-service';
 import { ThemeService } from '@theia/core/lib/browser/theming';
-import { MaybePromise } from '@theia/core/lib/common/types';
 import { WindowService } from '@theia/core/lib/browser/window/window-service';
 import { FileDialogService } from '@theia/filesystem/lib/browser/file-dialog/file-dialog-service';
 import { DisposableCollection } from '@theia/core/lib/common/disposable';
-import { FrontendApplicationStateService } from '@theia/core/lib/browser/frontend-application-state';
 import {
-  DialogProps,
-  PreferenceService,
-  PreferenceScope,
-  DialogError,
-  ReactWidget,
-} from '@theia/core/lib/browser';
-import { Index } from '../common/types';
-import {
-  CompilerWarnings,
   CompilerWarningLiterals,
-  ConfigService,
-  FileSystemExt,
   Network,
   ProxySettings,
-} from '../common/protocol';
-import { AbstractDialog } from './theia/dialogs/dialogs';
+} from '../../../common/protocol';
 import { nls } from '@theia/core/lib/common';
-
-const EDITOR_SETTING = 'editor';
-const FONT_SIZE_SETTING = `${EDITOR_SETTING}.fontSize`;
-const AUTO_SAVE_SETTING = `${EDITOR_SETTING}.autoSave`;
-const QUICK_SUGGESTIONS_SETTING = `${EDITOR_SETTING}.quickSuggestions`;
-const ARDUINO_SETTING = 'arduino';
-const WINDOW_SETTING = `${ARDUINO_SETTING}.window`;
-// const IDE_SETTING = `${ARDUINO_SETTING}.ide`;
-const COMPILE_SETTING = `${ARDUINO_SETTING}.compile`;
-const UPLOAD_SETTING = `${ARDUINO_SETTING}.upload`;
-const LANGUAGE_SETTING = `${ARDUINO_SETTING}.language`;
-const SKETCHBOOK_SETTING = `${ARDUINO_SETTING}.sketchbook`;
-const AUTO_SCALE_SETTING = `${WINDOW_SETTING}.autoScale`;
-const ZOOM_LEVEL_SETTING = `${WINDOW_SETTING}.zoomLevel`;
-// const AUTO_UPDATE_SETTING = `${IDE_SETTING}.autoUpdate`;
-const COMPILE_VERBOSE_SETTING = `${COMPILE_SETTING}.verbose`;
-const COMPILE_WARNINGS_SETTING = `${COMPILE_SETTING}.warnings`;
-const UPLOAD_VERBOSE_SETTING = `${UPLOAD_SETTING}.verbose`;
-const UPLOAD_VERIFY_SETTING = `${UPLOAD_SETTING}.verify`;
-const LANGUAGE_LOG_SETTING = `${LANGUAGE_SETTING}.log`;
-const SHOW_ALL_FILES_SETTING = `${SKETCHBOOK_SETTING}.showAllFiles`;
-
-export interface Settings extends Index {
-  editorFontSize: number; // `editor.fontSize`
-  themeId: string; // `workbench.colorTheme`
-  autoSave: 'on' | 'off'; // `editor.autoSave`
-  quickSuggestions: Record<'other' | 'comments' | 'strings', boolean>; // `editor.quickSuggestions`
-
-  autoScaleInterface: boolean; // `arduino.window.autoScale`
-  interfaceScale: number; // `arduino.window.zoomLevel` https://github.com/eclipse-theia/theia/issues/8751
-  checkForUpdates?: boolean; // `arduino.ide.autoUpdate`
-  verboseOnCompile: boolean; // `arduino.compile.verbose`
-  compilerWarnings: CompilerWarnings; // `arduino.compile.warnings`
-  verboseOnUpload: boolean; // `arduino.upload.verbose`
-  verifyAfterUpload: boolean; // `arduino.upload.verify`
-  enableLsLogs: boolean; // `arduino.language.log`
-  sketchbookShowAllFiles: boolean; // `arduino.sketchbook.showAllFiles`
-
-  sketchbookPath: string; // CLI
-  additionalUrls: string[]; // CLI
-  network: Network; // CLI
-}
-export namespace Settings {
-  export function belongsToCli<K extends keyof Settings>(key: K): boolean {
-    return key === 'sketchbookPath' || key === 'additionalUrls';
-  }
-}
-
-@injectable()
-export class SettingsService {
-  @inject(FileService)
-  protected readonly fileService: FileService;
-
-  @inject(FileSystemExt)
-  protected readonly fileSystemExt: FileSystemExt;
-
-  @inject(ConfigService)
-  protected readonly configService: ConfigService;
-
-  @inject(PreferenceService)
-  protected readonly preferenceService: PreferenceService;
-
-  @inject(FrontendApplicationStateService)
-  protected readonly appStateService: FrontendApplicationStateService;
-
-  protected readonly onDidChangeEmitter = new Emitter<Readonly<Settings>>();
-  readonly onDidChange = this.onDidChangeEmitter.event;
-
-  protected ready = new Deferred<void>();
-  protected _settings: Settings;
-
-  @postConstruct()
-  protected async init(): Promise<void> {
-    await this.appStateService.reachedState('ready'); // Hack for https://github.com/eclipse-theia/theia/issues/8993
-    const settings = await this.loadSettings();
-    this._settings = deepClone(settings);
-    this.ready.resolve();
-  }
-
-  protected async loadSettings(): Promise<Settings> {
-    await this.preferenceService.ready;
-    const [
-      editorFontSize,
-      themeId,
-      autoSave,
-      quickSuggestions,
-      autoScaleInterface,
-      interfaceScale,
-      // checkForUpdates,
-      verboseOnCompile,
-      compilerWarnings,
-      verboseOnUpload,
-      verifyAfterUpload,
-      enableLsLogs,
-      sketchbookShowAllFiles,
-      cliConfig,
-    ] = await Promise.all([
-      this.preferenceService.get<number>(FONT_SIZE_SETTING, 12),
-      this.preferenceService.get<string>(
-        'workbench.colorTheme',
-        'arduino-theme'
-      ),
-      this.preferenceService.get<'on' | 'off'>(AUTO_SAVE_SETTING, 'on'),
-      this.preferenceService.get<
-        Record<'other' | 'comments' | 'strings', boolean>
-      >(QUICK_SUGGESTIONS_SETTING, {
-        other: false,
-        comments: false,
-        strings: false,
-      }),
-      this.preferenceService.get<boolean>(AUTO_SCALE_SETTING, true),
-      this.preferenceService.get<number>(ZOOM_LEVEL_SETTING, 0),
-      // this.preferenceService.get<string>(AUTO_UPDATE_SETTING, true),
-      this.preferenceService.get<boolean>(COMPILE_VERBOSE_SETTING, true),
-      this.preferenceService.get<any>(COMPILE_WARNINGS_SETTING, 'None'),
-      this.preferenceService.get<boolean>(UPLOAD_VERBOSE_SETTING, true),
-      this.preferenceService.get<boolean>(UPLOAD_VERIFY_SETTING, true),
-      this.preferenceService.get<boolean>(LANGUAGE_LOG_SETTING, true),
-      this.preferenceService.get<boolean>(SHOW_ALL_FILES_SETTING, false),
-      this.configService.getConfiguration(),
-    ]);
-    const { additionalUrls, sketchDirUri, network } = cliConfig;
-    const sketchbookPath = await this.fileService.fsPath(new URI(sketchDirUri));
-    return {
-      editorFontSize,
-      themeId,
-      autoSave,
-      quickSuggestions,
-      autoScaleInterface,
-      interfaceScale,
-      // checkForUpdates,
-      verboseOnCompile,
-      compilerWarnings,
-      verboseOnUpload,
-      verifyAfterUpload,
-      enableLsLogs,
-      sketchbookShowAllFiles,
-      additionalUrls,
-      sketchbookPath,
-      network,
-    };
-  }
-
-  async settings(): Promise<Settings> {
-    await this.ready.promise;
-    return this._settings;
-  }
-
-  async update(settings: Settings, fireDidChange = false): Promise<void> {
-    await this.ready.promise;
-    for (const key of Object.keys(settings)) {
-      this._settings[key] = settings[key];
-    }
-    if (fireDidChange) {
-      this.onDidChangeEmitter.fire(this._settings);
-    }
-  }
-
-  async reset(): Promise<void> {
-    const settings = await this.loadSettings();
-    return this.update(settings, true);
-  }
-
-  async validate(
-    settings: MaybePromise<Settings> = this.settings()
-  ): Promise<string | true> {
-    try {
-      const { sketchbookPath, editorFontSize, themeId } = await settings;
-      const sketchbookDir = await this.fileSystemExt.getUri(sketchbookPath);
-      if (!(await this.fileService.exists(new URI(sketchbookDir)))) {
-        return nls.localize(
-          'arduino/preferences/invalid.sketchbook.location',
-          'Invalid sketchbook location: {0}',
-          sketchbookPath
-        );
-      }
-      if (editorFontSize <= 0) {
-        return nls.localize(
-          'arduino/preferences/invalid.editorFontSize',
-          'Invalid editor font size. It must be a positive integer.'
-        );
-      }
-      if (
-        !ThemeService.get()
-          .getThemes()
-          .find(({ id }) => id === themeId)
-      ) {
-        return nls.localize(
-          'arduino/preferences/invalid.theme',
-          'Invalid theme.'
-        );
-      }
-      return true;
-    } catch (err) {
-      if (err instanceof Error) {
-        return err.message;
-      }
-      return String(err);
-    }
-  }
-
-  async save(): Promise<string | true> {
-    await this.ready.promise;
-    const {
-      editorFontSize,
-      themeId,
-      autoSave,
-      quickSuggestions,
-      autoScaleInterface,
-      interfaceScale,
-      // checkForUpdates,
-      verboseOnCompile,
-      compilerWarnings,
-      verboseOnUpload,
-      verifyAfterUpload,
-      enableLsLogs,
-      sketchbookPath,
-      additionalUrls,
-      network,
-      sketchbookShowAllFiles,
-    } = this._settings;
-    const [config, sketchDirUri] = await Promise.all([
-      this.configService.getConfiguration(),
-      this.fileSystemExt.getUri(sketchbookPath),
-    ]);
-    (config as any).additionalUrls = additionalUrls;
-    (config as any).sketchDirUri = sketchDirUri;
-    (config as any).network = network;
-
-    await Promise.all([
-      this.preferenceService.set(
-        'editor.fontSize',
-        editorFontSize,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        'workbench.colorTheme',
-        themeId,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        'editor.autoSave',
-        autoSave,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        'editor.quickSuggestions',
-        quickSuggestions,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        AUTO_SCALE_SETTING,
-        autoScaleInterface,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        ZOOM_LEVEL_SETTING,
-        interfaceScale,
-        PreferenceScope.User
-      ),
-      // this.preferenceService.set(AUTO_UPDATE_SETTING, checkForUpdates, PreferenceScope.User),
-      this.preferenceService.set(
-        COMPILE_VERBOSE_SETTING,
-        verboseOnCompile,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        COMPILE_WARNINGS_SETTING,
-        compilerWarnings,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        UPLOAD_VERBOSE_SETTING,
-        verboseOnUpload,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        UPLOAD_VERIFY_SETTING,
-        verifyAfterUpload,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        LANGUAGE_LOG_SETTING,
-        enableLsLogs,
-        PreferenceScope.User
-      ),
-      this.preferenceService.set(
-        SHOW_ALL_FILES_SETTING,
-        sketchbookShowAllFiles,
-        PreferenceScope.User
-      ),
-      this.configService.setConfiguration(config),
-    ]);
-    this.onDidChangeEmitter.fire(this._settings);
-    return true;
-  }
-}
+import { Settings, SettingsService } from './settings';
+import { AdditionalUrlsDialog } from './settings-dialog';
+import { AsyncLocalizationProvider } from '@theia/core/lib/common/i18n/localization';
 
 export class SettingsComponent extends React.Component<
   SettingsComponent.Props,
@@ -440,6 +126,12 @@ export class SettingsComponent extends React.Component<
             </div>
             <div className="flex-line">
               {nls.localize(
+                'vscode/editorStatus/status.editor.mode',
+                'Language'
+              ) + ':'}
+            </div>
+            <div className="flex-line">
+              {nls.localize(
                 'arduino/preferences/showVerbose',
                 'Show verbose output during'
               )}
@@ -502,6 +194,27 @@ export class SettingsComponent extends React.Component<
                     </option>
                   ))}
               </select>
+            </div>
+            <div className="flex-line">
+              <select
+                className="theia-select"
+                value={this.state.currentLanguage}
+                onChange={this.languageDidChange}
+              >
+                {this.state.languages.map((label) => (
+                  <option key={label} value={label}>
+                    {label}
+                  </option>
+                ))}
+              </select>
+              <span style={{ marginLeft: '5px' }}>
+                (
+                {nls.localize(
+                  'vscode/extensionsActions/reloadRequired',
+                  'Reload required'
+                )}
+                )
+              </span>
             </div>
             <div className="flex-line">
               <label className="flex-line">
@@ -882,6 +595,13 @@ export class SettingsComponent extends React.Component<
     }
   };
 
+  protected languageDidChange = (
+    event: React.ChangeEvent<HTMLSelectElement>
+  ) => {
+    const selectedLanguage = event.target.value;
+    this.setState({ currentLanguage: selectedLanguage });
+  };
+
   protected compilerWarningsDidChange = (
     event: React.ChangeEvent<HTMLSelectElement>
   ) => {
@@ -1004,182 +724,7 @@ export namespace SettingsComponent {
     readonly fileService: FileService;
     readonly fileDialogService: FileDialogService;
     readonly windowService: WindowService;
+    readonly localizationProvider: AsyncLocalizationProvider;
   }
-  export type State = Settings;
-}
-
-@injectable()
-export class SettingsWidget extends ReactWidget {
-  @inject(SettingsService)
-  protected readonly settingsService: SettingsService;
-
-  @inject(FileService)
-  protected readonly fileService: FileService;
-
-  @inject(FileDialogService)
-  protected readonly fileDialogService: FileDialogService;
-
-  @inject(WindowService)
-  protected readonly windowService: WindowService;
-
-  protected render(): React.ReactNode {
-    return (
-      <SettingsComponent
-        settingsService={this.settingsService}
-        fileService={this.fileService}
-        fileDialogService={this.fileDialogService}
-        windowService={this.windowService}
-      />
-    );
-  }
-}
-
-@injectable()
-export class SettingsDialogProps extends DialogProps {}
-
-@injectable()
-export class SettingsDialog extends AbstractDialog<Promise<Settings>> {
-  @inject(SettingsService)
-  protected readonly settingsService: SettingsService;
-
-  @inject(SettingsWidget)
-  protected readonly widget: SettingsWidget;
-
-  constructor(
-    @inject(SettingsDialogProps)
-    protected readonly props: SettingsDialogProps
-  ) {
-    super(props);
-    this.contentNode.classList.add('arduino-settings-dialog');
-    this.appendCloseButton(
-      nls.localize('vscode/issueMainService/cancel', 'Cancel')
-    );
-    this.appendAcceptButton(nls.localize('vscode/issueMainService/ok', 'OK'));
-  }
-
-  @postConstruct()
-  protected init(): void {
-    this.toDispose.push(
-      this.settingsService.onDidChange(this.validate.bind(this))
-    );
-  }
-
-  protected async isValid(settings: Promise<Settings>): Promise<DialogError> {
-    const result = await this.settingsService.validate(settings);
-    if (typeof result === 'string') {
-      return result;
-    }
-    return '';
-  }
-
-  get value(): Promise<Settings> {
-    return this.settingsService.settings();
-  }
-
-  protected onAfterAttach(msg: Message): void {
-    if (this.widget.isAttached) {
-      Widget.detach(this.widget);
-    }
-    Widget.attach(this.widget, this.contentNode);
-    this.toDisposeOnDetach.push(
-      this.settingsService.onDidChange(() => this.update())
-    );
-    super.onAfterAttach(msg);
-    this.update();
-  }
-
-  protected onUpdateRequest(msg: Message) {
-    super.onUpdateRequest(msg);
-    this.widget.update();
-  }
-
-  protected onActivateRequest(msg: Message): void {
-    super.onActivateRequest(msg);
-
-    // calling settingsService.reset() in order to reload the settings from the preferenceService
-    // and update the UI including changes triggerd from the command palette
-    this.settingsService.reset();
-
-    this.widget.activate();
-  }
-}
-
-export class AdditionalUrlsDialog extends AbstractDialog<string[]> {
-  protected readonly textArea: HTMLTextAreaElement;
-
-  constructor(urls: string[], windowService: WindowService) {
-    super({
-      title: nls.localize(
-        'arduino/preferences/additionalManagerURLs',
-        'Additional Boards Manager URLs'
-      ),
-    });
-
-    this.contentNode.classList.add('additional-urls-dialog');
-
-    const description = document.createElement('div');
-    description.textContent = nls.localize(
-      'arduino/preferences/enterAdditionalURLs',
-      'Enter additional URLs, one for each row'
-    );
-    description.style.marginBottom = '5px';
-    this.contentNode.appendChild(description);
-
-    this.textArea = document.createElement('textarea');
-    this.textArea.className = 'theia-input';
-    this.textArea.setAttribute('style', 'flex: 0;');
-    this.textArea.value = urls
-      .filter((url) => url.trim())
-      .filter((url) => !!url)
-      .join('\n');
-    this.textArea.wrap = 'soft';
-    this.textArea.cols = 90;
-    this.textArea.rows = 5;
-    this.contentNode.appendChild(this.textArea);
-
-    const anchor = document.createElement('div');
-    anchor.classList.add('link');
-    anchor.textContent = nls.localize(
-      'arduino/preferences/unofficialBoardSupport',
-      'Click for a list of unofficial board support URLs'
-    );
-    anchor.style.marginTop = '5px';
-    anchor.style.cursor = 'pointer';
-    this.addEventListener(anchor, 'click', () =>
-      windowService.openNewWindow(
-        'https://github.com/arduino/Arduino/wiki/Unofficial-list-of-3rd-party-boards-support-urls',
-        { external: true }
-      )
-    );
-    this.contentNode.appendChild(anchor);
-
-    this.appendAcceptButton(nls.localize('vscode/issueMainService/ok', 'OK'));
-    this.appendCloseButton(
-      nls.localize('vscode/issueMainService/cancel', 'Cancel')
-    );
-  }
-
-  get value(): string[] {
-    return this.textArea.value
-      .split('\n')
-      .map((url) => url.trim())
-      .filter((url) => !!url);
-  }
-
-  protected onAfterAttach(message: Message): void {
-    super.onAfterAttach(message);
-    this.addUpdateListener(this.textArea, 'input');
-  }
-
-  protected onActivateRequest(message: Message): void {
-    super.onActivateRequest(message);
-    this.textArea.focus();
-  }
-
-  protected handleEnter(event: KeyboardEvent): boolean | void {
-    if (event.target instanceof HTMLInputElement) {
-      return super.handleEnter(event);
-    }
-    return false;
-  }
+  export type State = Settings & { languages: string[] };
 }
